refactor(dockview-core): move AdvancedDnD contract into core
Relocate IAdvancedDnDHost / IAdvancedDnDService into the core moduleContracts
file, like the other three modules. Core now imports zero types from any of the
four module implementation files — it references only the contracts — so each
module is fully decoupled from core. The service already encapsulates its whole
surface (preview overlay, group drag ghost, overlay-model resolution, onWill*
dispatch), so only the contract moved. No behaviour change; 1117 tests green.
