+++ packages/dockview-core/src/dockview/moduleContracts.ts
@@ -1,12 +1,13 @@
 /**
- * Contracts (service + host interfaces) for the pluggable feature modules whose
- * *implementations* are candidates to live outside core. Core references only
- * these interfaces — never a module's implementation — so a module's code can
- * be relocated without touching core. Keep this file implementation-free.
+ * Service + host interfaces for the pluggable feature modules. Core references
+ * only these interfaces — never a module's implementation — so each module is
+ * decoupled from core and independently testable / removable. Keep this file
+ * implementation-free.
  */
 import { IDisposable } from '../lifecycle';
 import { Event } from '../events';
-import { Position } from '../dnd/droptarget';
+import { DroptargetOverlayModel, Position } from '../dnd/droptarget';
+import { IDragGhostSpec } from '../dnd/backend';
 import { DockviewApi } from '../api/component.api';
 import { DockviewGroupPanel } from './dockviewGroupPanel';
 import { IDockviewPanel } from './dockviewPanel';
@@ -15,10 +16,17 @@ import { TabGroupColorPalette } from './tabGroupAccent';
 import { PopupService } from './components/popupService';
 import { DockviewComponentOptions } from './options';
 import { DockviewLayoutMutationEvent } from './dockviewComponent';
+import { DockviewWillDropEvent } from './dockviewGroupPanelModel';
 import {
+    GroupDragEvent,
+    TabDragEvent,
+} from './components/titlebar/tabsContainer';
+import {
+    DockviewGroupDropLocation,
     DockviewTabGroupChangeEvent,
     DockviewTabGroupPanelChangeEvent,
     DockviewTabGroupCollapsedChangeEvent,
+    DockviewWillShowOverlayLocationEvent,
 } from './events';
 
 // --- ContextMenu ---
@@ -107,3 +115,47 @@ export interface IAccessibilityHost {
 }
 
 export interface IAccessibilityService extends IDisposable {}
+
+// --- AdvancedDnD ---
+
+export interface IAdvancedDnDHost {
+    readonly options: DockviewComponentOptions;
+    readonly api: DockviewApi;
+    fireWillDragPanel(event: TabDragEvent): void;
+    fireWillDragGroup(event: GroupDragEvent): void;
+    fireWillDrop(event: DockviewWillDropEvent): void;
+    fireWillShowOverlay(event: DockviewWillShowOverlayLocationEvent): void;
+}
+
+export interface IAdvancedDnDService extends IDisposable {
+    dispatchWillDragPanel(event: TabDragEvent): void;
+    dispatchWillDragGroup(event: GroupDragEvent): void;
+    dispatchWillDrop(event: DockviewWillDropEvent): void;
+    dispatchWillShowOverlay(event: DockviewWillShowOverlayLocationEvent): void;
+    /**
+     * Resolve the custom group drag ghost from
+     * `createGroupDragGhostComponent`, or `undefined` when no factory is
+     * configured (the caller then renders the default chip). Returning
+     * `undefined` is also what happens when this module is absent.
+     */
+    buildGroupDragGhost(group: DockviewGroupPanel): IDragGhostSpec | undefined;
+    /**
+     * Resolve the app-supplied overlay model for a group drop target via the
+     * `dropOverlayModel` option, or `undefined` to keep the target's default.
+     */
+    resolveOverlayModel(
+        location: DockviewGroupDropLocation,
+        group?: DockviewGroupPanel
+    ): DroptargetOverlayModel | undefined;
+    /**
+     * Render the drop-preview overlay on a group at `position` — the same
+     * overlay a mouse drag shows — without a live drag. Returns a disposable
+     * that clears it. Used by keyboard docking so keyboard and mouse previews
+     * are identical (single source of truth). Commit the move via the public
+     * `api.moveGroupOrPanel({ to: { group, position } })`.
+     */
+    showPreviewOverlay(
+        group: DockviewGroupPanel,
+        position: Position
+    ): IDisposable;
+}
